# server/docs/swagger.yaml
definitions:
...
  api.NewsCreateReq:
    properties:
      content:
        type: string
      cover:
        type: string
      desc:
        type: string
      is_visible:
        description: 可见性
        type: boolean
      push_date:
        description: 发布日期 YYYY-MM-DD HH:mm:ss
        type: string
      recommend:
        description: 推荐等级 0 为不推荐，数值越大越靠前
        type: integer
      title:
        type: string
    required:
    - title
    type: object
  api.NewsCreateRes:
    properties:
      id:
        description: 新闻 ID
        type: integer
    required:
    - id
    type: object
  api.NewsInfoRes:
    properties:
      author_id:
        description: 作者
        type: integer
      content:
        description: 内容
        type: string
      cover:
        description: 封面
        type: string
      created_at:
        description: 创建时间
        type: string
      desc:
        description: 简介
        type: string
      id:
        description: ID
        type: integer
      is_visible:
        description: 可见性
        type: boolean
      push_date:
        description: 发布日期
        type: string
      recommend:
        description: 推荐等级
        type: integer
      title:
        description: 新闻标题
        type: string
      updated_at:
        description: 更新时间
        type: string
    required:
    - author_id
    - created_at
    - id
    - is_visible
    - recommend
    - title
    type: object
  api.NewsListReq:
    properties:
      current:
        minimum: 1
        type: integer
      keyword:
        description: 查询关键字
        type: string
      order_key:
        type: string
      order_type:
        enum:
        - ase
        - desc
        type: string
      page_size:
        minimum: 0
        type: integer
    type: object
  api.NewsListRes:
    properties:
      list:
        items:
          $ref: '#/definitions/model.News'
        type: array
      total:
        type: integer
    type: object
  api.NewsUpdateReq:
    properties:
      content:
        type: string
      cover:
        type: string
      desc:
        type: string
      id:
        description: 新闻 ID
        type: integer
      is_visible:
        description: 可见性
        type: boolean
      push_date:
        description: 发布日期 YYYY-MM-DD HH:mm:ss
        type: string
      recommend:
        description: 推荐等级 0 为不推荐，数值越大越靠前
        type: integer
      title:
        type: string
    required:
    - id
    - title
    type: object
...
  consts.CaptchaScenes:
    enum:
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-comments:
      CaptchaScenesForgetPassword: 密码重置
      CaptchaScenesRegister: 注册
      CaptchaScenesUpdateEmail: 更新邮箱
      CaptchaScenesUpdatePhone: 更新手机号
    x-enum-varnames:
    - CaptchaScenesRegister
    - CaptchaScenesForgetPassword
    - CaptchaScenesUpdatePhone
    - CaptchaScenesUpdateEmail
  consts.CaptchaType:
    enum:
    - 1
    - 2
    type: integer
    x-enum-comments:
      CaptchaTypeEmail: 邮箱
      CaptchaTypePhone: 手机号
    x-enum-varnames:
    - CaptchaTypePhone
    - CaptchaTypeEmail
  consts.UserStatus:
    enum:
    - 1
    - -1
    type: integer
    x-enum-comments:
      UserStatusLocked: 锁定
      UserStatusNormal: 正常
    x-enum-varnames:
    - UserStatusNormal
    - UserStatusLocked
  controller.CreateCaptchaBodyDto:
    properties:
      captcha_key:
        description: 图形验证码的key
        type: string
      captcha_value:
        description: 输入的图形验证码
        type: string
      scenes:
        allOf:
        - $ref: '#/definitions/consts.CaptchaScenes'
        description: 使用场景， 1-注册 2-忘记密码 3-修改手机 4-修改邮箱
      type:
        allOf:
        - $ref: '#/definitions/consts.CaptchaType'
        description: 验证码类型， 1-手机 2-邮箱
      value:
        description: 手机/邮箱账号
        type: string
    required:
    - captcha_key
    - captcha_value
    - scenes
    - type
    - value
    type: object
  controller.UpdateUserStatusBodyDto:
    properties:
      id:
        description: 用户 ID
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/consts.UserStatus'
        description: 状态 -1封禁 1-正常
        enum:
        - -1
        - 1
    required:
    - id
    - status
    type: object
...
  model.News:
    properties:
      author_id:
        description: 作者
        type: integer
      content:
        description: 内容
        type: string
      cover:
        description: 封面
        type: string
      created_at:
        description: 创建时间
        type: string
      desc:
        description: 简介
        type: string
      id:
        description: ID
        type: integer
      is_visible:
        description: 可见性
        type: boolean
      push_date:
        description: 发布日期
        type: string
      recommend:
        description: 推荐等级
        type: integer
      title:
        description: 新闻标题
        type: string
      updated_at:
        description: 更新时间
        type: string
    required:
    - author_id
    - created_at
    - id
    - is_visible
    - recommend
    - title
    type: object
  model.User:
    properties:
      age:
        type: integer
      avatar:
        type: string
      created_at:
        description: 创建时间
        type: string
      email:
        type: string
      id:
        description: ID
        type: integer
      name:
        type: string
      status:
        $ref: '#/definitions/consts.UserStatus'
      updated_at:
        description: 更新时间
        type: string
      username:
        type: string
      wx_open_id:
        type: string
      wx_union_id:
        type: string
    required:
    - created_at
    - email
    - id
    - status
    - username
    type: object
  resp.RList:
    properties:
      list:
        items: {}
        type: array
      total:
        type: integer
    type: object
  resp.Result:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  title: 男生自用 API 接口文档
  version: "1.0"
paths:
...
  /api/admin/c-user:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: 当前页
        in: query
        name: current
        type: number
      - default: 20
        description: 每页条数
        in: query
        name: page_size
        type: number
      - description: 需要排序的列
        in: query
        name: order_key
        type: string
      - description: 排序方式
        enum:
        - ase
        - desc
        in: query
        name: order_type
        type: string
      - description: 按名称或ID搜索
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            allOf:
            - $ref: '#/definitions/resp.Result'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/resp.RList'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/model.User'
                        type: array
                    type: object
              type: object
      summary: 用户列表
      tags:
      - 管理后台-C端用户管理
  /api/admin/news/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.NewsCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            allOf:
            - $ref: '#/definitions/resp.Result'
            - properties:
                data:
                  $ref: '#/definitions/api.NewsCreateRes'
              type: object
      summary: 新增新闻
      tags:
      - 管理后台-新闻
  /api/admin/news/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            $ref: '#/definitions/resp.Result'
      summary: 删除新闻
      tags:
      - 管理后台-新闻
  /api/admin/news/info:
    post:
      consumes:
      - application/json
      parameters:
      - description: Body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.NewsListReq'
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            allOf:
            - $ref: '#/definitions/resp.Result'
            - properties:
                data:
                  $ref: '#/definitions/api.NewsInfoRes'
              type: object
      summary: 新闻详情
      tags:
      - 管理后台-新闻
  /api/admin/news/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: Body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.NewsListReq'
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            allOf:
            - $ref: '#/definitions/resp.Result'
            - properties:
                data:
                  $ref: '#/definitions/api.NewsListRes'
              type: object
      summary: 新闻列表
      tags:
      - 管理后台-新闻
  /api/admin/news/update/info:
    post:
      consumes:
      - application/json
      parameters:
      - description: Body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.NewsUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            $ref: '#/definitions/resp.Result'
      summary: 修改新闻
      tags:
      - 管理后台-新闻
...
  /api/admin/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: Body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/controller.UpdateUserStatusBodyDto'
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            $ref: '#/definitions/resp.Result'
      summary: 修改用户状态
      tags:
      - 管理后台-C端用户管理
...
  /api/captcha:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/controller.CreateCaptchaBodyDto'
      produces:
      - application/json
      responses:
        "200":
          description: 验证码 ID
          schema:
            allOf:
            - $ref: '#/definitions/resp.Result'
            - properties:
                data:
                  type: integer
              type: object
      summary: 发送验证码
      tags:
      - 验证码
  /api/captcha/clear:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authentication token
        in: header
        name: X-Auth-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 请求结果
          schema:
            $ref: '#/definitions/resp.Result'
      summary: 清除过期验证码
      tags:
      - 验证码
...
  /api/captcha/image/{key}:
    get:
      consumes:
      - image/png
      parameters:
      - description: 图片验证码 key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 图片文件流
          schema:
            type: string
      summary: 获取验证码图片
      tags:
      - 验证码
...
swagger: "2.0"